Preserve high-resolution PPG samples until output stage

+++ src/components/ble/RawPpgService.h
@@ -16,7 +16,7 @@ namespace Pinetime {
       RawPpgService(NimbleController& nimble, Controllers::HeartRateController& heartRateController);
       void Init();
       int OnRawPpgRequested(uint16_t attributeHandle, ble_gatt_access_ctxt* context);
-      void OnNewRawPpgValue(uint16_t hrs, uint16_t als);
+      void OnNewRawPpgValue(uint32_t hrs, uint32_t als);
 
       void SubscribeNotification(uint16_t attributeHandle);
       void UnsubscribeNotification(uint16_t attributeHandle);
@@ -41,8 +41,8 @@ namespace Pinetime {
       struct ble_gatt_svc_def serviceDefinition[2];
 
       uint16_t rawPpgMeasurementHandle;
-      uint16_t lastHrs = 0;
-      uint16_t lastAls = 0;
+      uint32_t lastHrs = 0;
+      uint32_t lastAls = 0;
       uint32_t sampleCounter = 0;
       std::atomic_bool rawPpgMeasurementNotificationEnabled {false};
     };
